Use weibeld/ajv-cli container for GitHub Actions job
All the steps of the job will run in this container. This is implemented
by GitHub Actions by starting the container with a "tail -f /dev/null"
command (to keep it running), and then presumably executing the commands
of each step with "docker exec" in this container.

That means, all the steps run in the same container, and artefacts
created by one step (e.g. files) are available to subsequent steps.