Add "uri" format for fields containing URLs in JSON schemas
This validates that these fields contain valid URLs. This requires
loading the ajv-formats module [1] with the -c option [2] when invoking
ajv-cli because the formats are defined in this module (which has to be
installed as an npm package [3] in the Docker image).

[1] https://ajv.js.org/packages/ajv-formats.html
[2] https://ajv.js.org/packages/ajv-cli.html#c-custom-keywords-formats-definitions
[3] https://www.npmjs.com/package/ajv-formats

Temporarily force-pull Docker image in GitHub Actions jobs

For updating the used Docker image to the newest version (with the
ajv-formats package installed, but still having the same tag).

Remove force-pulling of Docker image