Move schema validation code to script and call from GitHub Actions
This allows to call the same code from a Git pre-commit hook, which
allows detecting errors earlier.

This also changes the organisation of the GitHub Actions workflow from
three jobs (one for each data file) to a single job that validates all
data files in a single step.

A possible approach is to organise further checks on the YAML files
(see #1, #3, #6) as separate jobs in the same workflow  (workflow may be
called 'Validate YAML files' with jobs called 'Schema', 'References',
'Duplicates', 'Linting').