Fix bug in scripts related to while loop in sub-shell
Previously, the 'failed' variable was set in a while loop used as
'files | while; do ... done;'. In this case, the while loop runs in a
sub-shell because it is the right-hand side of a pipe. That means that
any modifications to the 'failed' variable in the while loop were lost
outside of the loop. This caused the scripts to terminate with a zero
exit code even if one of the validations failed.

This was a bug introduced in fbb004c7c2d40de7910fc2dafac234fc0b74edc4.

Now, the 'failed' status is saved in a file rather than in a variable
which guarantees that the status is seen identically from all execution
environments.