Reorganise GitHub Actions workflow scripts
Each script runs itself autonomously in a Docker container: when a
script starts, it checks whether it's running in a container, and if
not, starts the container and launches itself in this container. This
makes the executions of the scripts uniform (no more executions on the
host system with manually satisfied dependencies) and also keeps the
scripts self-contained (no helper scripts necessary, scripts ensure
they're running in a container themselves).

Besides this, the following changes have been performed:

- Change name of scripts directory from 'workflows' to 'scripts'.
- Change name of GitHub Actions workflow from 'validate-data-files' to
  'validate'.
- Adapt workflow definition of 'validate' workflow to not use any
  container settings anymore but just execute the scripts (since now the
  scripts take care themselves of the container execution environment).
- Adapt the pre-commit hook to simply execute the workflow scripts
  (which now results in exactly the same execution environment because
  the scripts run in containers).